requirements.txt
# ========= Core scientific stack =========
numpy>=1.24,<2.0          # TF2.10 は NumPy 2非対応なので <2 に固定
pandas>=2.0,<2.3
scipy>=1.10,<1.13

# ========= Plotting / Visualization =========
matplotlib>=3.7,<3.9
seaborn>=0.13,<0.14
plotly>=5.0,<6.0
japanize-matplotlib>=1.1,<1.2  # 日本語グラフ必要なら

# ========= Time-series & econometrics =========
statsmodels>=0.14,<0.15   # 回帰・時系列モデル
arch>=6.3,<6.4            # GARCH系（もう入れてるやつ）
pymc>=5.0,<6.0            # ベイズ推定（PyMC v5系）
pystan>=3.9,<4.0          # PyMCのバックエンド用に

# ========= Machine Learning / Deep Learning =========
scikit-learn==1.4.2    # 古いのでもいいけど1.3〜ならOK

# PyTorch は GPU 版を手動インストール済なら、ここではバージョン縛らないかコメント化
# （GPUSOROBANで cu117 のホイール使うなら requirements からは外しておいて、
#   別途 `pip install torch==2.0.1+cu117 torchvision==0.15.2+cu117 ...` とかが無難）
# torch==2.0.1

# ========= Finance / Data sources =========
yfinance>=0.2,<0.3        # Yahoo Finance から株価など
ccxt>=4.3,<5.0            # 仮想通貨取引所API用

# QuantLibはネイティブ拡張＋依存ライブラリが要るので注意。
# SOROBANの環境に `libboost` などが入っていれば:
# QuantLib==1.36
# 入らなければ、最初はコメントアウトしておいて後で追加インストールでもOK。

# ========= Infra / ユーティリティ =========
pybitflyer
python-dotenv>=1.0,<2.0   # .env 読み込み
# optional (for text analysis / pretty plots)
# wordcloud>=1.9,<2.0
# fugashi[unidic-lite]>=1.2,<2.0
